requirements.txt
# ======================
# BugHunter Pro v7.0 - Realistic Implementation Requirements
# ======================

# Core Dependencies (Required)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0

# Phase 1: Core Infrastructure
aiohttp>=3.9.0              # Async HTTP
uvloop>=0.19.0              # Fast async event loop (Unix/Mac)
pyyaml>=6.0.0               # YAML configuration
apscheduler>=3.10.0         # Scheduled tasks (CVE sync)

# Phase 2: CVE & Exploit Database
# All use standard requests library

# Machine Learning Features
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Advanced & Reliability
tenacity>=8.2.0
redis>=5.0.0                # Optional: Distributed caching

# Security & Crypto Analysis
cryptography>=41.0.0
pyOpenSSL>=23.3.0           # TLS/SSL analysis

# HTML Parsing Enhancements
html5lib>=1.1

# Performance Monitoring
psutil>=5.9.0

# PDF Reporting
reportlab>=4.0.0

# Optional Advanced Features (Install separately)
# playwright>=1.40.0        # Headless browser (Phase 3)
# androguard>=3.4.0         # APK analysis (Phase 4)
# web3>=6.11.0              # Blockchain scanning (Phase 4)

# Development Dependencies (Do not install in production)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
